+++ src/option.ts
@@ -5,6 +5,7 @@ export type BuilderOption = {
   addAsBeforeColumnAlias?: 'always'|'never'
   addPublicSchemaName?: 'always'|'never'|'when other schema mentioned'
   addTableName?: 'always'|'when two tables or more' //TODO: add more options like 'when needed' and 'when conflict'
+  addAsBeforeTableAlias?: 'always'|'never'
 }
 
 const defaultOption: BuilderOption = {
@@ -14,6 +15,7 @@ const defaultOption: BuilderOption = {
   addAsBeforeColumnAlias: 'always',
   addPublicSchemaName: 'never',
   addTableName: 'when two tables or more',
+  addAsBeforeTableAlias: 'always',
 }
 Object.freeze(defaultOption)
 
@@ -25,5 +27,6 @@ export function fillUndefinedOptionsWithDefault(option: BuilderOption): BuilderO
   result.addAsBeforeColumnAlias = option.addAsBeforeColumnAlias ?? defaultOption.addAsBeforeColumnAlias
   result.addPublicSchemaName = option.addPublicSchemaName ?? defaultOption.addPublicSchemaName
   result.addTableName = option.addTableName ?? defaultOption.addTableName
+  result.addAsBeforeTableAlias = option.addAsBeforeTableAlias ?? defaultOption.addAsBeforeTableAlias
   return result
 }