+++ test/database.ts
@@ -27,6 +27,12 @@ export const database = new Database({
             col8: new BooleanColumn({ name: 'col8' }),
           },
         }),
+        table2: new Table({
+          name: 'table2',
+          columns: {
+            col1: new TextColumn({ name: 'col1' }),
+          },
+        }),
       },
     }),
     schema1: new Schema({
@@ -34,6 +40,12 @@ export const database = new Database({
       tables: {
         table1: new Table({
           name: 'table1',
+          columns: {
+            col1: new TextColumn({ name: 'col1' }),
+          },
+        }),
+        table2: new Table({
+          name: 'table2',
           columns: {
             col1: new TextColumn({ name: 'col1' }),
           },