+++ test/options.test.ts
@@ -1,9 +1,11 @@
 import { Builder } from '../src'
 import { database } from './database'
 //Alias
-const table = database.s.public.t.testTable
+const testTable = database.s.public.t.testTable
 const col1 = database.s.public.t.testTable.c.col1
 const col2 = database.s.public.t.testTable.c.col2
+const table2 = database.s.public.t.table2
+const table2col1 = database.s.public.t.table2.c.col1
 const table1 = database.s.schema1.t.table1
 
 describe('test Options', () => {
@@ -14,7 +16,7 @@ describe('test Options', () => {
     it('Produces [SELECT 1 FROM "testTable"] without semicolon', () => {
       const actual = sqlWithoutSemicolon
         .select(1)
-        .from(table)
+        .from(testTable)
         .getSQL()
 
       expect(actual).toEqual('SELECT 1 FROM "testTable"')
@@ -23,7 +25,7 @@ describe('test Options', () => {
     it('Produces [SELECT 1 FROM "testTable"] without semicolon;', () => {
       const actual = sqlWithSemicolon
         .select(1)
-        .from(table)
+        .from(testTable)
         .getSQL()
 
       expect(actual).toEqual('SELECT 1 FROM "testTable";')
@@ -32,7 +34,7 @@ describe('test Options', () => {
     it('Produces [SELECT 1 FROM "testTable"] without semicolon; (default)', () => {
       const actual = sqlDefault
         .select(1)
-        .from(table)
+        .from(testTable)
         .getSQL()
 
       expect(actual).toEqual('SELECT 1 FROM "testTable";')
@@ -54,7 +56,7 @@ describe('test Options', () => {
     it('Produces [SELECT "col1" FROM "testTable" ORDER BY "col1" ASC;] option(always)', () => {
       const actual = sqlAlways
         .select(col1)
-        .from(table)
+        .from(testTable)
         .orderBy(col1)
         .getSQL()
 
@@ -64,7 +66,7 @@ describe('test Options', () => {
     it('Produces [SELECT "col1" FROM "testTable" ORDER BY "col1";] option(never)', () => {
       const actual = sqlNever
         .select(col1)
-        .from(table)
+        .from(testTable)
         .orderBy(col1)
         .getSQL()
 
@@ -74,7 +76,7 @@ describe('test Options', () => {
     it('Produces [SELECT "col1" FROM "testTable" ORDER BY "col1", "col2" DESC;] option(never)', () => {
       const actual = sqlNever
         .select(col1)
-        .from(table)
+        .from(testTable)
         .orderBy(col1, col2.desc)
         .getSQL()
 
@@ -84,7 +86,7 @@ describe('test Options', () => {
     it('Produces [SELECT "col1" FROM "testTable" ORDER BY "col1";] option(never) even asc mentioned', () => {
       const actual = sqlNever
         .select(col1)
-        .from(table)
+        .from(testTable)
         .orderBy(col1.asc)
         .getSQL()
 
@@ -94,7 +96,7 @@ describe('test Options', () => {
     it('Produces [SELECT "col1" FROM "testTable" ORDER BY "col1" ASC;] option(when mentioned)', () => {
       const actual = sqlWhenMentioned
         .select(col1)
-        .from(table)
+        .from(testTable)
         .orderBy(col1.asc)
         .getSQL()
 
@@ -104,7 +106,7 @@ describe('test Options', () => {
     it('Produces [SELECT "col1" FROM "testTable" ORDER BY "col1";] option(when mentioned)', () => {
       const actual = sqlWhenMentioned
         .select(col1)
-        .from(table)
+        .from(testTable)
         .orderBy(col1)
         .getSQL()
 
@@ -114,7 +116,7 @@ describe('test Options', () => {
     it('Produces [SELECT "col1" FROM "testTable" ORDER BY "col1";] option(Default)', () => {
       const actual = sqlDefault
         .select(col1)
-        .from(table)
+        .from(testTable)
         .orderBy(col1)
         .getSQL()
 
@@ -124,7 +126,7 @@ describe('test Options', () => {
     it('Produces [SELECT "col1" FROM "testTable" ORDER BY "col1" ASC;] option(Default)', () => {
       const actual = sqlDefault
         .select(col1)
-        .from(table)
+        .from(testTable)
         .orderBy(col1.asc)
         .getSQL()
 
@@ -147,7 +149,7 @@ describe('test Options', () => {
     it('Produces [SELECT "col1" FROM "testTable" ORDER BY "col1" NULLS LAST;] option(always)', () => {
       const actual = sqlAlways
         .select(col1)
-        .from(table)
+        .from(testTable)
         .orderBy(col1)
         .getSQL()
 
@@ -157,7 +159,7 @@ describe('test Options', () => {
     it('Produces [SELECT "col1" FROM "testTable" ORDER BY "col1";] option(never)', () => {
       const actual = sqlNever
         .select(col1)
-        .from(table)
+        .from(testTable)
         .orderBy(col1)
         .getSQL()
 
@@ -167,7 +169,7 @@ describe('test Options', () => {
     it('Produces [SELECT "col1" FROM "testTable" ORDER BY "col1" NULLS FIRST ;] option(never)', () => {
       const actual = sqlNever
         .select(col1)
-        .from(table)
+        .from(testTable)
         .orderBy(col1.nullsFirst)
         .getSQL()
 
@@ -177,7 +179,7 @@ describe('test Options', () => {
     it('Produces [SELECT "col1" FROM "testTable" ORDER BY "col1";] option(never) even nulls last mentioned', () => {
       const actual = sqlNever
         .select(col1)
-        .from(table)
+        .from(testTable)
         .orderBy(col1.nullsLast)
         .getSQL()
 
@@ -187,7 +189,7 @@ describe('test Options', () => {
     it('Produces [SELECT "col1" FROM "testTable" ORDER BY "col1" ASC;] option(when mentioned)', () => {
       const actual = sqlWhenMentioned
         .select(col1)
-        .from(table)
+        .from(testTable)
         .orderBy(col1.asc)
         .getSQL()
 
@@ -197,7 +199,7 @@ describe('test Options', () => {
     it('Produces [SELECT "col1" FROM "testTable" ORDER BY "col1";] option(when mentioned)', () => {
       const actual = sqlWhenMentioned
         .select(col1)
-        .from(table)
+        .from(testTable)
         .orderBy(col1)
         .getSQL()
 
@@ -207,7 +209,7 @@ describe('test Options', () => {
     it('Produces [SELECT "col1" FROM "testTable" ORDER BY "col1";] option(Default)', () => {
       const actual = sqlDefault
         .select(col1)
-        .from(table)
+        .from(testTable)
         .orderBy(col1)
         .getSQL()
 
@@ -217,7 +219,7 @@ describe('test Options', () => {
     it('Produces [SELECT "col1" FROM "testTable" ORDER BY "col1" NULLS LAST;] option(Default)', () => {
       const actual = sqlDefault
         .select(col1)
-        .from(table)
+        .from(testTable)
         .orderBy(col1.nullsLast)
         .getSQL()
 
@@ -239,7 +241,7 @@ describe('test Options', () => {
     it('Produces [SELECT "col1" AS "C1" FROM "testTable";] option(always)', () => {
       const actual = sqlAlways
         .select(col1.as('C1'))
-        .from(table)
+        .from(testTable)
         .getSQL()
 
       expect(actual).toEqual('SELECT "col1" AS "C1" FROM "testTable";')
@@ -248,7 +250,7 @@ describe('test Options', () => {
     it('Produces [SELECT "col1" FROM "testTable";] option(never)', () => {
       const actual = sqlNever
         .select(col1.as('C1'))
-        .from(table)
+        .from(testTable)
         .getSQL()
 
       expect(actual).toEqual('SELECT "col1" "C1" FROM "testTable";')
@@ -257,7 +259,7 @@ describe('test Options', () => {
     it('Produces [SELECT "col1" AS "C1" FROM "testTable";] option(default)', () => {
       const actual = sqlDefault
         .select(col1.as('C1'))
-        .from(table)
+        .from(testTable)
         .getSQL()
 
       expect(actual).toEqual('SELECT "col1" AS "C1" FROM "testTable";')
@@ -265,78 +267,93 @@ describe('test Options', () => {
   })
 
   describe('test addPublicSchemaName Option', () => {
-    const sqlAlways = new Builder(database, { addPublicSchemaName: 'always' })
-    const sqlWhen = new Builder(database, { addPublicSchemaName: 'when other schema mentioned' })
-    const sqlNever = new Builder(database, { addPublicSchemaName: 'never' })
-    const sqlDefault = new Builder(database)
+    describe('Option: always', () => {
+      const sql = new Builder(database, { addPublicSchemaName: 'always' })
+      afterEach(() => { sql.cleanUp() })
 
-    afterEach(() => {
-      sqlAlways.cleanUp()
-      sqlWhen.cleanUp()
-      sqlNever.cleanUp()
-      sqlDefault.cleanUp()
+      it('Produces [SELECT "col1" FROM "public"."testTable";]', () => {
+        const actual = sql
+          .select(col1)
+          .from(testTable)
+          .getSQL()
+
+        expect(actual).toEqual('SELECT "col1" FROM "public"."testTable";')
+      })
     })
 
-    it('Produces [SELECT "col1" FROM "public"."testTable";] option(always)', () => {
-      const actual = sqlAlways
-        .select(col1)
-        .from(table)
-        .getSQL()
+    describe('Option: when other schema mentioned', () => {
+      const sql = new Builder(database, { addPublicSchemaName: 'when other schema mentioned' })
+      afterEach(() => { sql.cleanUp() })
 
-      expect(actual).toEqual('SELECT "col1" FROM "public"."testTable";')
+      it('Produces [SELECT "col2", "col1" FROM "public"."testTable", "schema1"."table1";', () => {
+        const actual = sql
+          .select(testTable.c.col2, table1.c.col1)
+          .from(testTable, table1)
+          .getSQL()
+
+        expect(actual).toEqual('SELECT "testTable"."col2", "table1"."col1" FROM "public"."testTable", "schema1"."table1";')
+      })
     })
 
-    it('Produces [SELECT "col1" FROM "testTable";] option(when)', () => {
-      //TODO: add another table to this test when builder takes more than one table
-      const actual = sqlWhen
-        .select(col1)
-        .from(table)
-        .getSQL()
+    describe('Option: never', () => {
+      const sql = new Builder(database, { addPublicSchemaName: 'never' })
+      afterEach(() => { sql.cleanUp() })
 
-      expect(actual).toEqual('SELECT "col1" FROM "testTable";')
+      it('Produces [SELECT "public"."table2"."col1", "schema1"."table2"."col1" FROM "table2", "schema1"."table2";]', () => {
+        const actual = sql
+          .select(
+            database.s.public.t.table2.c.col1,
+            database.s.schema1.t.table2.c.col1,
+          )
+          .from(
+            database.s.public.t.table2,
+            database.s.schema1.t.table2,
+          )
+          .getSQL()
+
+        expect(actual).toEqual('SELECT "public"."table2"."col1", "schema1"."table2"."col1" FROM "table2", "schema1"."table2";')
+      })
+
+      it('Produces [SELECT "col1" FROM "schema1"."table1";]', () => {
+        const actual = sql
+          .select(table1.c.col1)
+          .from(table1)
+          .getSQL()
+
+        expect(actual).toEqual('SELECT "col1" FROM "schema1"."table1";')
+      })
     })
+    describe('Option: default', () => {
+      const sql = new Builder(database)
+      afterEach(() => { sql.cleanUp() })
 
-    it('Produces [SELECT "col1" FROM "testTable";] option(never)', () => {
-      //TODO: add another table to this test when builder takes more than one table
-      const actual = sqlNever
-        .select(col1)
-        .from(table)
-        .getSQL()
+      it('Produces [SELECT "public"."table2"."col1", "schema1"."table2"."col1" FROM "table2", "schema1"."table2";]', () => {
+        const actual = sql
+          .select(
+            database.s.public.t.table2.c.col1,
+            database.s.schema1.t.table2.c.col1,
+          )
+          .from(
+            database.s.public.t.table2,
+            database.s.schema1.t.table2,
+          )
+          .getSQL()
 
-      expect(actual).toEqual('SELECT "col1" FROM "testTable";')
-    })
+        expect(actual).toEqual('SELECT "public"."table2"."col1", "schema1"."table2"."col1" FROM "table2", "schema1"."table2";')
+      })
 
+      it('Produces [SELECT "col1" FROM "schema1"."table1";]', () => {
+        const actual = sql
+          .select(table1.c.col1)
+          .from(table1)
+          .getSQL()
 
-    it('Produces [SELECT "col1" FROM "schema1"."table1";] option(never)', () => {
-      const actual = sqlNever
-        .select(table1.c.col1)
-        .from(table1)
-        .getSQL()
-
-      expect(actual).toEqual('SELECT "col1" FROM "schema1"."table1";')
-    })
-
-    it('Produces [SELECT "col1" FROM "testTable";] option(default)', () => {
-      //TODO: add another table to this test when builder takes more than one table
-      const actual = sqlDefault
-        .select(col1)
-        .from(table)
-        .getSQL()
-
-      expect(actual).toEqual('SELECT "col1" FROM "testTable";')
-    })
-
-    it('Produces [SELECT "col1" FROM "schema1"."table1";] option(default)', () => {
-      const actual = sqlDefault
-        .select(table1.c.col1)
-        .from(table1)
-        .getSQL()
-
-      expect(actual).toEqual('SELECT "col1" FROM "schema1"."table1";')
+        expect(actual).toEqual('SELECT "col1" FROM "schema1"."table1";')
+      })
     })
   })
 
-  describe('test addPublicSchemaName Option', () => {
+  describe('test addTableName Option', () => {
     const sqlAlways = new Builder(database, { addTableName: 'always' })
     const sqlWhen = new Builder(database, { addTableName: 'when two tables or more' })
     const sqlDefault = new Builder(database)
@@ -350,30 +367,87 @@ describe('test Options', () => {
     it('Produces [SELECT "testTable"."col1" FROM "testTable";] option(always)', () => {
       const actual = sqlAlways
         .select(col1)
-        .from(table)
+        .from(testTable)
         .getSQL()
 
       expect(actual).toEqual('SELECT "testTable"."col1" FROM "testTable";')
     })
 
     it('Produces [SELECT "col1" FROM "testTable";] option(when)', () => {
-      //TODO: add another table to this test when builder takes more than one table
       const actual = sqlWhen
         .select(col1)
-        .from(table)
+        .from(testTable)
         .getSQL()
 
       expect(actual).toEqual('SELECT "col1" FROM "testTable";')
     })
 
+    it('Produces [SELECT "testTable"."col1", "table2"."col1" FROM "testTable", "table2";] option(when)', () => {
+      const actual = sqlWhen
+        .select(col1, table2col1)
+        .from(testTable, table2)
+        .getSQL()
+
+      expect(actual).toEqual('SELECT "testTable"."col1", "table2"."col1" FROM "testTable", "table2";')
+    })
+
     it('Produces [SELECT "col1" FROM "testTable";] option(default)', () => {
-      //TODO: add another table to this test when builder takes more than one table
       const actual = sqlDefault
         .select(col1)
-        .from(table)
+        .from(testTable)
         .getSQL()
 
       expect(actual).toEqual('SELECT "col1" FROM "testTable";')
+    })
+
+    it('Produces [SELECT "testTable"."col1", "table2"."col1" FROM "testTable", "table2";] option(default)', () => {
+      const actual = sqlDefault
+        .select(col1, table2col1)
+        .from(testTable, table2)
+        .getSQL()
+
+      expect(actual).toEqual('SELECT "testTable"."col1", "table2"."col1" FROM "testTable", "table2";')
+    })
+  })
+
+  describe('test addAsBeforeTableAlias Option', () => {
+    describe('Option: always', () => {
+      const sqlAlways = new Builder(database, { addAsBeforeTableAlias: 'always' })
+      afterEach(() => { sqlAlways.cleanUp() })
+      it('Produces [SELECT "testTable"."col1" FROM "testTable" AS "TEST Table";]', () => {
+        const actual = sqlAlways
+          .select(col1)
+          .from(testTable.as('TEST Table'))
+          .getSQL()
+
+        expect(actual).toEqual('SELECT "col1" FROM "testTable" AS "TEST Table";')
+      })
+    })
+
+    describe('Option: never', () => {
+      const sqlNever = new Builder(database, { addAsBeforeTableAlias: 'never' })
+      afterEach(() => { sqlNever.cleanUp() })
+      it('Produces [SELECT "testTable"."col1" FROM "testTable" "TEST Table";]', () => {
+        const actual = sqlNever
+          .select(col1)
+          .from(testTable.as('TEST Table'))
+          .getSQL()
+
+        expect(actual).toEqual('SELECT "col1" FROM "testTable" "TEST Table";')
+      })
+    })
+
+    describe('Option: default', () => {
+      const sqlDefault = new Builder(database)
+      afterEach(() => { sqlDefault.cleanUp() })
+      it('Produces [SELECT "testTable"."col1" FROM "testTable" AS "TEST Table";]', () => {
+        const actual = sqlDefault
+          .select(col1)
+          .from(testTable.as('TEST Table'))
+          .getSQL()
+
+        expect(actual).toEqual('SELECT "col1" FROM "testTable" AS "TEST Table";')
+      })
     })
   })
 })