+++ CHANGELOG.md
@@ -1,6 +1,16 @@
 # Changelog
 <!-- https://keepachangelog.com/en/1.0.0/ -->
 
+## [0.11.0]  2022-04-10
+### Changes
+- table data builder change to array of FromItemInfo
+
+### Added
+- From step can take more than one table
+- Builder option addAsBeforeTableAlias to Builder Option
+- AliasedTable class, table can be aliased
+- CrossJoin Step
+
 ## [0.10.1]  2022-04-08
 ### Changes
 - Rename "column?" to "col?" in unit test
@@ -12,14 +22,14 @@
 - Move Operand class into Operand.ts file
 - Rename models to types and move it inside models folder
 - Rename file select.ts to SelectItemInfo.ts
-- Add IStatementGiver Interface
-- Add option addPublicSchemaName to Builder Option
-- Add Builder data to Table getStmt
-- Add option addTableName to Builder Option
-- Add Builder data to Column getStmt
 
 ### Added
 - Having step can contain aggregate function condition
+- IStatementGiver Interface
+- option addPublicSchemaName to Builder Option
+- Builder data to Table getStmt
+- option addTableName to Builder Option
+- Builder data to Column getStmt
 
 ## [0.10.0]  2022-03-26
 ### Changes